requirements.txt
django>=4.0
psycopg2-binary>=2.8
djangorestframework>=3.7
pydantic
python-dotenv
